Removed calculation of tiles for nonmoving unit
For each step in a movement phase units need to update their visibility of other units, both to account for the moving unit and units revealed by it getting out of the way. Only the moving unit has the possibility of revealing new tiles, which means that we can save some computational effort by using the new FOV calculation's update-units-only-feature.