Bug when using mods like Expanded Inventory (overlaps body)
Unable to equip/unequip some inventory slots that overlap the "change armor" button in InventoryState because the mouse event is fired on the Inventory, but also on the button with some weird effects. Solution - "consume" the mouseClick Action by replacing when appropriate with a mouseMove Action (SDLEvent change) having deltax and y 0. Fixed an issue when unable to unload weapon and Ufopaedia opened for both equipped item and armor.